% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factory_functions.R
\name{factory_function_utils}
\alias{factory_function_utils}
\alias{shift_support}
\alias{factory_cdf_randomly_stopped_sum}
\title{Function-building utilities}
\usage{
shift_support(args, distribution_function, shift_args, env = parent.frame())

factory_cdf_randomly_stopped_sum(
  args,
  conditional_cdf,
  duration_pmf,
  conditional_cdf_argument_transformations,
  env = parent.frame()
)
}
\arguments{
\item{args}{arguments passed by the user to the created function}

\item{distribution_function}{(...)}

\item{shift_args}{string describing transformations to inputs of the
distribution function, usually a 'rigid shift' indicating a different support}

\item{env}{environment in which to evaluate the new function}
}
\description{
Function-building utilities
}
