% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_shifted.R
\name{geom_shifted}
\alias{geom_shifted}
\alias{dgeom_shifted_manual}
\alias{dgeom_shifted_programmatic}
\title{Geometric Distribution Supported on Positive Integers}
\usage{
dgeom_shifted_manual(x, prob, log = FALSE)

dgeom_shifted_programmatic(x, prob, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles representing the number of trials in a sequence of Bernoulli trials before success occurs.}

\item{prob}{probability of success in each trial. \code{0 < prob <= 1}.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\description{
Geometric Distribution Supported on Positive Integers
}
