% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gse.R
\name{cdf_gse}
\alias{cdf_gse}
\title{The joint distribution of a geometrically-stopped sums of IID exponential
RVs}
\usage{
cdf_gse(duration, q, rate, prob)
}
\description{
The joint distribution of a geometrically-stopped sums of IID exponential
RVs
}
